pyproject.toml
[project]
name = "mailflow"
version = "0.3.0"
description = "Smart email processing for mutt with similarity learning and optional LLM classification"
readme = "README.md"
authors = [
    { name = "Juan Reyero", email = "juan@juanreyero.com" }
]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "click>=8.2.1",
    "gnureadline>=8.2.13",
    "google-api-python-client>=2.178.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "playwright>=1.54.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "slack_sdk>=3.33.4",
    "pypdf>=5.1.0",
    "llmring",
]

[project.scripts]
mailflow = "mailflow.cli:cli"
mailflow-setup = "mailflow.setup:main"
mailflow-mcp = "mailflow.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest-asyncio>=0.26.0",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
line_length = 99

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "C4",   # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B", "SIM"]  # Allow assert statements and complexity in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
norecursedirs = ["llmring"]

[tool.uv.sources]
llmring = { path = "../llmring-all/llmring" }
